bank-data class
